Finish query rule

+++ x-pack/plugins/security_solution/server/lib/detection_engine/rule_types/field_maps/rules.ts
@@ -102,7 +102,7 @@ export const rulesFieldMap = {
     required: true,
   },
   'kibana.alert.rule.threshold.value': {
-    type: 'long',
+    type: 'float', // TODO: should be 'long'
     array: true,
     required: true,
   },
